feat: done refactor comments

+++ src/Applications/use_case/_test/DeleteThreadCommentUseCase.test.js
@@ -3,7 +3,7 @@ const ThreadRepository = require('../../../Domains/threads/ThreadRepository');
 const CommentRepository = require('../../../Domains/comments/CommentRepository');
 
 describe('DeleteThreadCommentUseCase', () => {
-  it('should throw error when comment not found', async () => {
+  it('should throw error when thread not found', async () => {
     const mockRepository = new ThreadRepository();
     const mockCommentRepository = new CommentRepository();
     const payload = {
@@ -12,18 +12,16 @@ describe('DeleteThreadCommentUseCase', () => {
       threadId: 'thread-123',
     };
 
-    mockRepository.getThreadById = jest.fn()
-      .mockImplementation(() => Promise.resolve({
-        rowCount: 0,
-      }));
+    mockRepository.verifyAvailabilityThreadById = jest.fn()
+      .mockImplementation(() => Promise.reject(new Error('GET_THREAD.NO_THREAD_FOUND')));
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
       threadRepository: mockRepository,
       commentRepository: mockCommentRepository,
     });
     await expect(deleteThreadCommentUseCase.execute(payload)).rejects.toThrowError('GET_THREAD.NO_THREAD_FOUND');
-    expect(mockRepository.getThreadById).toBeCalledWith('thread-123');
+    expect(mockRepository.verifyAvailabilityThreadById).toBeCalledWith('thread-123');
   });
-  it('should throw error when comment not found', async () => {
+  it('should throw error when comment not found in thread', async () => {
     const mockRepository = new ThreadRepository();
     const mockCommentRepository = new CommentRepository();
     const payload = {
@@ -32,19 +30,17 @@ describe('DeleteThreadCommentUseCase', () => {
       threadId: 'thread-123',
     };
 
-    mockRepository.getThreadById = jest.fn()
-      .mockImplementation(() => Promise.resolve({
-        rowCount: 1,
-      }));
-    mockCommentRepository.getCommentByIdAndThreadId = jest.fn()
-      .mockImplementation(() => Promise.resolve({ rowCount: 0 }));
+    mockRepository.verifyAvailabilityThreadById = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyAvailabilityCommentByIdAndThreadId = jest.fn()
+      .mockImplementation(() => Promise.reject(new Error('GET_THREAD_COMMENT.NO_THREAD_COMMENT_FOUND')));
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
       threadRepository: mockRepository,
       commentRepository: mockCommentRepository,
     });
     await expect(deleteThreadCommentUseCase.execute(payload)).rejects.toThrowError('GET_THREAD_COMMENT.NO_THREAD_COMMENT_FOUND');
-    expect(mockRepository.getThreadById).toBeCalledWith('thread-123');
-    expect(mockCommentRepository.getCommentByIdAndThreadId).toBeCalledWith('comment-123', 'thread-123');
+    expect(mockRepository.verifyAvailabilityThreadById).toBeCalledWith('thread-123');
+    expect(mockCommentRepository.verifyAvailabilityCommentByIdAndThreadId).toBeCalledWith('comment-123', 'thread-123');
   });
   it('should throw error when user try delete another user comment', async () => {
     const mockRepository = new ThreadRepository();
@@ -55,34 +51,20 @@ describe('DeleteThreadCommentUseCase', () => {
       threadId: 'thread-123',
     };
 
-    mockRepository.getThreadById = jest.fn()
-      .mockImplementation(() => Promise.resolve({
-        rowCount: 1,
-        rows: [
-          {
-            id: 'thread-123',
-            user_id: 'user-124',
-          },
-        ],
-      }));
-    mockCommentRepository.getCommentByIdAndThreadId = jest.fn()
-      .mockImplementation(() => Promise.resolve({
-        rowCount: 1,
-        rows: [
-          {
-            id: 'comment-123',
-            thread_id: 'thread-123',
-            user_id: 'user-124',
-          },
-        ],
-      }));
+    mockRepository.verifyAvailabilityThreadById = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyAvailabilityCommentByIdAndThreadId = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyOwnerCommentByIdAndUserId = jest.fn()
+      .mockImplementation(() => Promise.reject(new Error('VERIFY_COMMENT_OWNER.ACCESS_FORBIDEN')));
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
       threadRepository: mockRepository,
       commentRepository: mockCommentRepository,
     });
-    await expect(deleteThreadCommentUseCase.execute(payload)).rejects.toThrowError('DELETE_THREAD_COMMENT.ACCESS_FORBIDEN');
-    expect(mockRepository.getThreadById).toBeCalledWith('thread-123');
-    expect(mockCommentRepository.getCommentByIdAndThreadId).toBeCalledWith('comment-123', 'thread-123');
+    await expect(deleteThreadCommentUseCase.execute(payload)).rejects.toThrowError('VERIFY_COMMENT_OWNER.ACCESS_FORBIDEN');
+    expect(mockRepository.verifyAvailabilityThreadById).toBeCalledWith('thread-123');
+    expect(mockCommentRepository.verifyAvailabilityCommentByIdAndThreadId).toBeCalledWith('comment-123', 'thread-123');
+    expect(mockCommentRepository.verifyOwnerCommentByIdAndUserId).toBeCalledWith('comment-123', 'user-123');
   });
   it('should delete comment thread correcly', async () => {
     const mockRepository = new ThreadRepository();
@@ -93,54 +75,28 @@ describe('DeleteThreadCommentUseCase', () => {
       userId: 'user-123',
       threadId: 'thread-123',
     };
-    const expected = {
-      rowCount: 1,
-      rows: [
-        {
-          id: 'comment-123',
-          user_id: 'user-123',
-        },
-      ],
-    };
 
-    mockRepository.getThreadById = jest.fn()
-      .mockImplementation(() => Promise.resolve({
-        rowCount: 1,
-        rows: [
-          {
-            id: 'thread-123',
-            user_id: 'user-123',
-          },
-        ],
-      }));
-    mockCommentRepository.getCommentByIdAndThreadId = jest.fn()
-      .mockImplementation(() => Promise.resolve({
-        rowCount: 1,
-        rows: [
-          {
-            id: 'comment-123',
-            thread_id: 'thread-123',
-            user_id: 'user-123',
-          },
-        ],
-      }));
+    mockRepository.verifyAvailabilityThreadById = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyAvailabilityCommentByIdAndThreadId = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyOwnerCommentByIdAndUserId = jest.fn()
+      .mockImplementation(() => Promise.resolve());
     mockCommentRepository.deleteCommentById = jest.fn()
-      .mockImplementation(() => Promise.resolve({
-        rowCount: 1,
-        rows: [
-          {
-            id: 'comment-123',
-            user_id: 'user-123',
-          },
-        ],
-      }));
+      .mockImplementation(() => Promise.resolve());
+    const spyDelete = jest.spyOn(mockCommentRepository, 'deleteCommentById');
     const deleteThreadCommentUseCase = new DeleteThreadCommentUseCase({
       threadRepository: mockRepository,
       commentRepository: mockCommentRepository,
     });
-    await expect(deleteThreadCommentUseCase.execute(payload)).resolves.toStrictEqual(expected);
-    expect(mockRepository.getThreadById).toBeCalledWith('thread-123');
-    expect(mockCommentRepository.getCommentByIdAndThreadId).toBeCalledWith('comment-123', 'thread-123');
-    expect(mockCommentRepository.deleteCommentById).toBeCalledWith('comment-123');
+    await deleteThreadCommentUseCase.execute(payload);
+    expect(mockRepository.verifyAvailabilityThreadById).toBeCalledWith(payload.threadId);
+    expect(mockCommentRepository.verifyAvailabilityCommentByIdAndThreadId)
+      .toBeCalledWith(payload.commentId, payload.threadId);
+    expect(mockCommentRepository.verifyOwnerCommentByIdAndUserId)
+      .toBeCalledWith(payload.commentId, payload.userId);
+    expect(mockCommentRepository.deleteCommentById).toBeCalledWith(payload.commentId);
+    expect(spyDelete).toBeCalledTimes(1);
+    expect(spyDelete).toBeCalledWith(payload.commentId);
   });
 });
